+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - Save downloaded files with URL-decoded filename
 
+- Don't pass deltachat-internal info messages to chatwoot
+
 ## 1.0.0
 
 - Release initial version